NewDriver file tidying
